requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
fastapi==0.115.0
h11==0.14.0
idna==3.10
joblib==1.4.2
lxml==5.3.0
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pikepdf==9.2.1
pillow==10.4.0
pip-tools==7.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
wheel==0.44.0
wrapt==1.16.0
